Removed namespaces
