+++ source/Utilities/FeatureDetectorUtility/include/FeatureDetectorUtilities.hh
@@ -63,7 +63,7 @@ class FeatureDetectorConfig: public crl::multisense::system::SecondaryAppConfig 
 
     private:
 
-        wire::FeatureDetectorConfigParams mConfigItems;
+        FeatureDetectorConfigParams mConfigItems;
 
     public:
         /**
@@ -210,8 +210,47 @@ public:
  * @return A crl::multisense::Status indicating if the callback deregistration
  * succeeded or failed
  */
-Status secondaryAppDataExtract(feature_detector::Header &header, const secondary_app::Header &orig);
+Status secondaryAppDataExtract(feature_detector::Header &header, const secondary_app::Header &orig)
+{
+  using namespace crl::multisense::details;
 
+  utility::BufferStreamReader stream(reinterpret_cast<const uint8_t*>(orig.secondaryAppDataP), orig.secondaryAppDataLength);
+  FeatureDetector featureDetector(stream, FeatureDetector::VERSION);
+
+  utility::BufferStreamReader metaStream(reinterpret_cast<const uint8_t *>(orig.secondaryAppMetadataP), orig.secondaryAppMetadataLength);
+  FeatureDetectorMeta _meta(metaStream, FeatureDetectorMeta::VERSION);
+
+
+  header.source         = featureDetector.source;
+  header.frameId        = _meta.frameId;
+  header.timeSeconds    = _meta.timeSeconds;
+  header.timeNanoSeconds= _meta.timeNanoSeconds;
+  header.ptpNanoSeconds = _meta.ptpNanoSeconds;
+  header.octaveWidth    = _meta.octaveWidth;
+  header.octaveHeight   = _meta.octaveHeight;
+  header.numOctaves     = _meta.numOctaves;
+  header.scaleFactor    = _meta.scaleFactor;
+  header.motionStatus   = _meta.motionStatus;
+  header.averageXMotion = _meta.averageXMotion;
+  header.averageYMotion = _meta.averageYMotion;
+  header.numFeatures    = featureDetector.numFeatures;
+  header.numDescriptors = featureDetector.numDescriptors;
+
+  const size_t startDescriptor=featureDetector.numFeatures*sizeof(Feature);
+
+  uint8_t * dataP = reinterpret_cast<uint8_t *>(featureDetector.dataP);
+  for (size_t i = 0; i < featureDetector.numFeatures; i++) {
+      feature_detector::Feature f = *reinterpret_cast<feature_detector::Feature *>(dataP + (i * sizeof(Feature)));
+      header.features->push_back(f);
+  }
+
+  for (size_t j = 0;j < featureDetector.numDescriptors; j++) {
+      feature_detector::Descriptor d = *reinterpret_cast<feature_detector::Descriptor *>(dataP + (startDescriptor + (j * sizeof(Descriptor))));
+      header.descriptors->push_back(d);
+  }
+
+  return Status_Ok;
+}
 
 } // namespace util
